+++ orcid-core/src/main/java/org/orcid/core/manager/read_only/impl/PersonDetailsManagerReadOnlyImpl.java
@@ -1,6 +1,7 @@
 package org.orcid.core.manager.read_only.impl;
 
 import java.util.ArrayList;
+import java.util.List;
 import java.util.stream.Collectors;
 
 import org.orcid.core.common.manager.EmailDomainManager;
@@ -14,6 +15,7 @@ import org.orcid.core.manager.read_only.ProfileKeywordManagerReadOnly;
 import org.orcid.core.manager.read_only.RecordNameManagerReadOnly;
 import org.orcid.core.manager.read_only.ResearcherUrlManagerReadOnly;
 import org.orcid.core.utils.SourceEntityUtils;
+import org.orcid.jaxb.model.common_v2.Source;
 import org.orcid.jaxb.model.common_v2.Visibility;
 import org.orcid.jaxb.model.record_v2.Address;
 import org.orcid.jaxb.model.record_v2.Addresses;
@@ -31,6 +33,8 @@ import org.orcid.jaxb.model.record_v2.PersonExternalIdentifiers;
 import org.orcid.jaxb.model.record_v2.ResearcherUrl;
 import org.orcid.jaxb.model.record_v2.ResearcherUrls;
 import org.orcid.persistence.jpa.entities.EmailDomainEntity;
+
+import liquibase.repackaged.org.apache.commons.lang3.StringUtils;
 
 import javax.annotation.Resource;
 
@@ -138,7 +142,6 @@ public class PersonDetailsManagerReadOnlyImpl extends ManagerReadOnlyBaseImpl im
         if (emails.getEmails() != null) {
             Emails filteredEmails = new Emails();
             filteredEmails.setEmails(new ArrayList<Email>(emails.getEmails().stream().filter(e -> e.isVerified()).collect(Collectors.toList())));
-            processProfessionalEmails(filteredEmails);
             person.setEmails(filteredEmails);
         }
         return person;
@@ -197,23 +200,9 @@ public class PersonDetailsManagerReadOnlyImpl extends ManagerReadOnlyBaseImpl im
         if (emails.getEmails() != null) {
             Emails filteredEmails = new Emails();
             filteredEmails.setEmails(new ArrayList<Email>(emails.getEmails()));
-            processProfessionalEmails(filteredEmails);
             person.setEmails(filteredEmails);
         }
 
         return person;
-    }
-
-    private void processProfessionalEmails(Emails emails) {
-        for (Email email : emails.getEmails()) {
-            if (email.isVerified()) {
-                String domain = email.getEmail().split("@")[1];
-                EmailDomainEntity domainInfo = emailDomainManager.findByEmailDomain(domain);
-                // Set appropriate source name and source id for professional emails
-                if (domainInfo != null && domainInfo.getCategory().equals(EmailDomainEntity.DomainCategory.PROFESSIONAL)) {
-                    email.setSource(sourceEntityUtils.convertEmailSourceToOrcidValidator(email.getSource()));
-                }
-            }
-        }
     }
 }