+++ orcid-persistence/src/main/java/org/orcid/persistence/dao/impl/PublicApiDailyRateLimitDaoImpl.java
@@ -114,6 +114,19 @@ public class PublicApiDailyRateLimitDaoImpl extends GenericDaoImpl<PublicApiDail
         return;
     }
 
+    @Override
+    @Transactional
+    public int cleanup(int daysToKeep) {
+        if(daysToKeep <= 1) {
+            throw new IllegalArgumentException("daysToKeep must be greater than 1");
+        }
+        LocalDate lastDayToKeep = LocalDate.now().minusDays(daysToKeep);
+        LOG.info("About to remove public_api_daily_rate_limit older than {}", lastDayToKeep);
+        Query query = entityManager.createNativeQuery("DELETE FROM public_api_daily_rate_limit WHERE date_created < :lastDayToKeep");
+        query.setParameter("lastDayToKeep", lastDayToKeep);
+        return query.executeUpdate();
+    }
+
     private static String logQueryWithParams(String baseQuery, Map<String, Object> params) {
         for (Map.Entry<String, Object> entry : params.entrySet()) {
             String paramPlaceholder = ":" + entry.getKey();